backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlmodel>=0.0.22
aiosqlite==0.20.0
pydantic>=2.9.2,<3.0.0
python-dotenv==1.0.1
APScheduler==3.10.4
croniter==2.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

